Fix checking of key/value pairs.
Original problem was a typo, but apparently you "can't add a new key into hash during iteration" so we need to create a new hash of string-keyed conclusion values. The keys need to be strings b/c the user-input key may be something like "111000" that will get interpreted as an integer and so can't be converted to a symbol.